Chapter18/Exercise18.03/bookr/requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-configurations==2.2
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
gunicorn==20.0.4
psycopg2==2.8.5